Refactor duplicated quote verifying fn
