+++ crates/shared/Cargo.toml
@@ -23,6 +23,7 @@ lazy_static={ version="1.5.0", features=["spin_no_std"] }
 hex-literal="0.4.1"
 sp-core    ={ version="29.0.0", default-features=false }
 subxt      ={ version="0.35.3", default-features=false, optional=true }
+tdx-quote  ="0.0.3"
 
 [features]
 default    =["std"]
@@ -31,3 +32,5 @@ wasm       =["codec/std", "scale-info/std", "serde/std", "sp-std/std"]
 wasm-no-std=["sp-runtime"]
 user-native=["dep:subxt", "subxt/native"]
 user-wasm  =["dep:subxt", "subxt/web"]
+# Enables non-mock TDX quote verification
+production =[]
