+++ pallets/attestation/Cargo.toml
@@ -61,4 +61,4 @@ std=[
 ]
 try-runtime=['frame-support/try-runtime']
 # When enabled, use real PCK certificate chain verification
-production=[]
+production=['entropy-shared/production']
