+++ pallets/staking/src/lib.rs
@@ -58,8 +58,9 @@ use sp_staking::SessionIndex;
 #[frame_support::pallet]
 pub mod pallet {
     use entropy_shared::{
-        QuoteContext, ValidatorInfo, VerifyQuoteError, X25519PublicKey, MAX_SIGNERS,
-        PREGENERATED_NETWORK_VERIFYING_KEY, VERIFICATION_KEY_LENGTH,
+        attestation::{AttestationHandler, QuoteContext, VerifyQuoteError},
+        ValidatorInfo, X25519PublicKey, MAX_SIGNERS, PREGENERATED_NETWORK_VERIFYING_KEY,
+        VERIFICATION_KEY_LENGTH,
     };
     use frame_support::{
         dispatch::{DispatchResult, DispatchResultWithPostInfo},
@@ -105,7 +106,7 @@ pub mod pallet {
         type MaxEndpointLength: Get<u32>;
 
         /// The handler to use when issuing and verifying attestations.
-        type AttestationHandler: entropy_shared::AttestationHandler<Self::AccountId>;
+        type AttestationHandler: AttestationHandler<Self::AccountId>;
     }
 
     /// Endpoint where a threshold server can be reached at
@@ -433,7 +434,9 @@ pub mod pallet {
                 if let Some(server_info) = maybe_server_info {
                     // Before we modify the `server_info`, we want to check that the validator is
                     // still running TDX hardware.
-                    <T::AttestationHandler as entropy_shared::AttestationHandler<_>>::verify_quote(
+                    <T::AttestationHandler as entropy_shared::attestation::AttestationHandler<
+                        _,
+                    >>::verify_quote(
                         &server_info.tss_account.clone(),
                         server_info.x25519_public_key,
                         quote,
@@ -497,7 +500,7 @@ pub mod pallet {
                         // Before we modify the `server_info`, we want to check that the validator is
                         // still running TDX hardware.
                         let provisioning_certification_key =
-                            <T::AttestationHandler as entropy_shared::AttestationHandler<_>>::verify_quote(
+                            <T::AttestationHandler as entropy_shared::attestation::AttestationHandler<_>>::verify_quote(
                                 &tss_account.clone(),
                                 x25519_public_key,
                                 quote,
@@ -632,7 +635,7 @@ pub mod pallet {
             );
 
             let provisioning_certification_key =
-                <T::AttestationHandler as entropy_shared::AttestationHandler<_>>::verify_quote(
+                <T::AttestationHandler as entropy_shared::attestation::AttestationHandler<_>>::verify_quote(
                     &joining_server_info.tss_account.clone(),
                     joining_server_info.x25519_public_key,
                     quote,
